pyproject.toml
[tool.poetry]
name = "newsflow"
version = "0.1.0"
description = "A GitHub Actions powered RSS newsreader that fetches and updates RSS feeds to a JSON file for frontend display."
authors = ["spezifisch <spezifisch+git@gmail.com>"]
license = "AGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
feedparser = "^6.0.11"
toml = "^0.10.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
